The beginning
I have started with an empty html document to create my new website.
Compared to the one I have now, it will be a lot cleaner and to the
point. I want it to be pleasing not only to the eye when viewing on any
device but also pleasing and easy to use and navigate.